(#4) refactor: fix package location
